Controller tests, some validation improvement and bug fixes

+++ src/main/java/com/codyperry/reading_tracker/dto/UpdateProgressResponse.java
@@ -1,4 +1,4 @@
 package com.codyperry.reading_tracker.dto;
 
-public record UpdateProgressResponse(long id, String name, String author, double percentComplete) {
+public record UpdateProgressResponse(long id, String name, String author, int percentComplete) {
 }
